requirements.txt
setuptools>=69
wheel>=0.43
-e .
PyYAML>=6.0
polars>=1.30.0
dpkt>=1.9.8
tqdm>=4.66
pytest>=8.0
